+++ src/tokio/child/unix.rs
@@ -57,7 +57,7 @@ impl ChildImp {
 		self.inner
 	}
 
-	pub(super) fn signal_imp(&mut self, sig: Signal) -> Result<()> {
+	pub(super) fn signal_imp(&self, sig: Signal) -> Result<()> {
 		killpg(self.pgid, sig).map_err(Error::from)
 	}
 
@@ -133,7 +133,7 @@ impl ChildImp {
 }
 
 impl crate::UnixChildExt for ChildImp {
-	fn signal(&mut self, sig: Signal) -> Result<()> {
+	fn signal(&self, sig: Signal) -> Result<()> {
 		self.signal_imp(sig)
 	}
 }